% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_method_freq.analysis.R
\name{freq.analysis}
\alias{freq.analysis}
\alias{freq.analysis,kRp.taggedText-method}
\title{Analyze word frequencies}
\usage{
freq.analysis(txt.file, ...)

\S4method{freq.analysis}{kRp.taggedText}(
  txt.file,
  corp.freq = NULL,
  desc.stat = TRUE,
  corp.rm.class = "nonpunct",
  corp.rm.tag = c(),
  tfidf = TRUE,
  as.feature = FALSE
)
}
\arguments{
\item{txt.file}{An object of class \code{\link[koRpus:kRp.tagged-class]{kRp.tagged}}.}

\item{...}{Additional options to be passed through to the function defined with \code{tagger}.}

\item{corp.freq}{An object of class \code{\link[koRpus:kRp.corp.freq-class]{kRp.corp.freq}}.}

\item{desc.stat}{Logical, whether a descriptive statistical analysis should be performed.}

\item{corp.rm.class}{A character vector with word classes which should be ignored for frequency analysis. The default value
\code{"nonpunct"} has special meaning and will cause the result of
\code{kRp.POS.tags(lang, tags=c("punct","sentc"), list.classes=TRUE)} to be used.}

\item{corp.rm.tag}{A character vector with POS tags which should be ignored for frequency analysis.}

\item{tfidf}{Logical,
      whether the term frequency--inverse document frequency statistic (tf-idf) should be computed. Requires
\code{corp.freq} to provide appropriate idf values for the types in \code{txt.file}. Missing idf values will result in \code{NA}.}

\item{as.feature}{Logical,
      whether the output should be just the analysis results or the input object with
the results added as a feature. Use \code{\link[koRpus:corpusFreq]{corpusFreq}}
to get the results from such an aggregated object.}
}
\value{
Depending on \code{as.feature},
      either an object of class \code{\link[koRpus:kRp.txt.freq-class]{kRp.txt.freq}},
   or an object of class \code{\link[koRpus:kRp.tagged-class]{kRp.tagged}} with the added feature \code{freq} containing it.
}
\description{
The function \code{freq.analysis} analyzes texts regarding frequencies of tokens,
      word classes etc.
}
\details{
The easiest way to see what kinds of analyses are done is probably to look at the slot description of \code{\link[koRpus:kRp.txt.freq-class]{kRp.txt.freq}}.
}
\examples{
\dontrun{
freq.analysis(tagged.text, corp.freq=my.LCC.data)
}
}
\seealso{
\code{\link[koRpus:get.kRp.env]{get.kRp.env}},
      \code{\link[koRpus:kRp.tagged-class]{kRp.tagged}},
   \code{\link[koRpus:kRp.corp.freq-class]{kRp.corp.freq}}
}
\keyword{misc}
