% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions_readability.R
\name{fucks}
\alias{fucks}
\title{Readability: Fucks' Stilcharakteristik}
\usage{
fucks(txt.file, ...)
}
\arguments{
\item{txt.file}{Either an object of class \code{\link[koRpus:kRp.text-class]{kRp.text}},
      a character vector which must be be
a valid path to a file containing the text to be analyzed,
      or a list of text features. If the latter, calculation
is done by \code{\link[koRpus:readability.num]{readability.num}}.}

\item{...}{Further valid options for the main function,
      see \code{\link[koRpus:readability]{readability}} for details.}
}
\value{
An object of class \code{\link[koRpus:kRp.readability-class]{kRp.readability}}.
}
\description{
This is just a convenient wrapper function for \code{\link[koRpus:readability]{readability}}.
}
\details{
Calculates Fucks' Stilcharakteristik ("characteristics of style"; Fucks, 1955,
      as cited in Briest, 1974).
In contrast to \code{\link[koRpus:readability]{readability}},
      which by default calculates all possible indices,
this function will only calculate the index value.
}
\examples{
\dontrun{
  fucks(tagged.text)
}
}
\references{
Briest, W. (1974). Kann man Verständlichkeit messen? \emph{Zeitschrift für Phonetik,
      Sprachwissenschaft und Kommunikationsforschung}, 27, 543--563.
}
\keyword{readability}
