% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_method_filterByClass.R,
%   R/02_method_pasteText.R, R/02_method_readTagged.R,
%   R/02_method_textTransform.R, R/kRp.text.analysis.R
\name{koRpus-deprecated}
\alias{koRpus-deprecated}
\alias{kRp.filter.wclass}
\alias{kRp.text.paste}
\alias{read.tagged}
\alias{kRp.text.transform}
\alias{kRp.text.analysis}
\title{Deprecated functions}
\usage{
kRp.filter.wclass(...)

kRp.text.paste(...)

read.tagged(...)

kRp.text.transform(...)

kRp.text.analysis(...)
}
\arguments{
\item{...}{Any option to be ignored.}
}
\description{
These functions will be removed soon and should no longer ne used.

This was basically a wrapper for \code{treetag()},
     \code{freq.analysis()} and \code{lex.div()}.
Please use those directly for equivalent results and more control over your workflow.
}
