% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koRpus-package.R
\docType{package}
\name{koRpus-package}
\alias{koRpus}
\alias{koRpus-package}
\title{\packageTitle{koRpus}}
\description{
\packageDescription{koRpus}
}
\details{
The DESCRIPTION file:
\tabular{ll}{
Package: \tab koRpus\cr
Type: \tab Package\cr
Version: \tab 0.13-9\cr
Date: \tab 2025-12-01\cr
Depends: \tab R (>= 3.0.0),sylly (>= 0.1-7)\cr
Enhances: \tab rkward\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab https://reaktanz.de/?c=hacking&s=koRpus\cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://reaktanz.de/?c=hacking&s=koRpus}
  \item Report bugs at \url{https://github.com/unDocUMeantIt/koRpus/issues}
}

}
\author{
\packageAuthor{koRpus}

Maintainer: \packageMaintainer{koRpus}
}
