\docType{methods}
\name{taggedText}
\alias{describe}
\alias{describe<-}
\alias{describe<-,kRp.taggedText-method}
\alias{describe,kRp.taggedText-method}
\alias{describe<-,-methods}
\alias{describe,-methods}
\alias{is.taggedText}
\alias{language}
\alias{language<-}
\alias{language<-,kRp.taggedText-method}
\alias{language,kRp.taggedText-method}
\alias{language<-,-methods}
\alias{language,-methods}
\alias{taggedText}
\alias{taggedText<-}
\alias{taggedText<-,kRp.taggedText-method}
\alias{taggedText,kRp.taggedText-method}
\alias{taggedText<-,-methods}
\alias{taggedText,-methods}
\title{Getter/setter methods for koRpus objects}
\usage{
  is.taggedText(obj)
}
\arguments{
  \item{obj}{An arbitrary \code{R} object.}
}
\description{
  These methods should be used to get or set values of
  tagged text objects generated by koRpus functions like
  \code{treetag()} or \code{tokenize()}.
}
\details{
  \itemize{ \item{\code{taggedText()} }{returns the
  \code{TT.res} slot.} \item{\code{describe()} }{returns
  the \code{desc} slot.} \item{\code{language()} }{returns
  the \code{lang} slot.} }
}
\examples{
\dontrun{
taggedText(tagged.txt)
}
}

