% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_method_kRp.taggedText.R
\docType{methods}
\name{taggedText}
\alias{taggedText}
\alias{taggedText,kRp.taggedText-method}
\alias{taggedText,-methods}
\alias{taggedText<-}
\alias{taggedText<-,kRp.taggedText-method}
\alias{taggedText<-,-methods}
\alias{hasFeature}
\alias{hasFeature,kRp.taggedText-method}
\alias{hasFeature,-methods}
\alias{hasFeature<-}
\alias{hasFeature<-,kRp.taggedText-method}
\alias{hasFeature<-,-methods}
\alias{feature}
\alias{feature,kRp.taggedText-method}
\alias{feature,-methods}
\alias{feature<-}
\alias{feature<-,kRp.taggedText-method}
\alias{feature<-,-methods}
\alias{corpusReadability}
\alias{corpusReadability,kRp.taggedText-method}
\alias{corpusReadability,-methods}
\alias{corpusReadability<-}
\alias{corpusReadability<-,kRp.taggedText-method}
\alias{corpusReadability<-,-methods}
\alias{corpusHyphen}
\alias{corpusHyphen,kRp.taggedText-method}
\alias{corpusHyphen,-methods}
\alias{corpusHyphen<-}
\alias{corpusHyphen<-,kRp.taggedText-method}
\alias{corpusHyphen<-,-methods}
\alias{corpusLexDiv}
\alias{corpusLexDiv,kRp.taggedText-method}
\alias{corpusLexDiv,-methods}
\alias{corpusLexDiv<-}
\alias{corpusLexDiv<-,kRp.taggedText-method}
\alias{corpusLexDiv<-,-methods}
\alias{corpusFreq}
\alias{corpusFreq,kRp.taggedText-method}
\alias{corpusFreq,-methods}
\alias{corpusFreq<-}
\alias{corpusFreq<-,kRp.taggedText-method}
\alias{corpusFreq<-,-methods}
\alias{corpusCorpFreq}
\alias{corpusCorpFreq,kRp.taggedText-method}
\alias{corpusCorpFreq,-methods}
\alias{corpusCorpFreq<-}
\alias{corpusCorpFreq<-,kRp.taggedText-method}
\alias{corpusCorpFreq<-,-methods}
\alias{corpusStopwords}
\alias{corpusStopwords,kRp.taggedText-method}
\alias{corpusStopwords,-methods}
\alias{corpusStopwords<-}
\alias{corpusStopwords<-,kRp.taggedText-method}
\alias{corpusStopwords<-,-methods}
\alias{[,kRp.taggedText,ANY,ANY,ANY-method}
\alias{[,-methods}
\alias{[,kRp.taggedText,ANY,ANY-method}
\alias{[<-,kRp.taggedText,ANY,ANY,ANY-method}
\alias{[<-,-methods}
\alias{[[,kRp.taggedText-method}
\alias{[[,-methods}
\alias{[[,kRp.taggedText,ANY-method}
\alias{[[<-,kRp.taggedText-method}
\alias{[[<-,-methods}
\alias{[[<-,kRp.taggedText,ANY,ANY-method}
\alias{describe,kRp.taggedText-method}
\alias{describe,-methods}
\alias{describe<-,kRp.taggedText-method}
\alias{describe<-,-methods}
\alias{language,kRp.taggedText-method}
\alias{language,-methods}
\alias{language<-,kRp.taggedText-method}
\alias{language<-,-methods}
\alias{diffText}
\alias{diffText,kRp.taggedText-method}
\alias{diffText,-methods}
\alias{diffText<-}
\alias{diffText<-,kRp.taggedText-method}
\alias{diffText<-,-methods}
\alias{originalText}
\alias{originalText,kRp.taggedText-method}
\alias{originalText,-methods}
\alias{is.taggedText}
\alias{fixObject}
\alias{fixObject,kRp.taggedText-method}
\alias{fixObject,-methods}
\alias{tif_as_tokens_df}
\alias{tif_as_tokens_df,kRp.taggedText-method}
\alias{tif_as_tokens_df,-methods}
\title{Getter/setter methods for koRpus objects}
\usage{
taggedText(obj, add.desc = FALSE, doc_id = FALSE)

\S4method{taggedText}{kRp.taggedText}(obj, add.desc = FALSE, doc_id = FALSE)

taggedText(obj) <- value

\S4method{taggedText}{kRp.taggedText}(obj) <- value

hasFeature(obj, feature = NULL, ...)

\S4method{hasFeature}{kRp.taggedText}(obj, feature = NULL)

hasFeature(obj, feature) <- value

\S4method{hasFeature}{kRp.taggedText}(obj, feature) <- value

feature(obj, feature, ...)

\S4method{feature}{kRp.taggedText}(obj, feature)

feature(obj, feature) <- value

\S4method{feature}{kRp.taggedText}(obj, feature) <- value

corpusReadability(obj, ...)

\S4method{corpusReadability}{kRp.taggedText}(obj)

corpusReadability(obj) <- value

\S4method{corpusReadability}{kRp.taggedText}(obj) <- value

corpusHyphen(obj, ...)

\S4method{corpusHyphen}{kRp.taggedText}(obj)

corpusHyphen(obj) <- value

\S4method{corpusHyphen}{kRp.taggedText}(obj) <- value

corpusLexDiv(obj, ...)

\S4method{corpusLexDiv}{kRp.taggedText}(obj)

corpusLexDiv(obj) <- value

\S4method{corpusLexDiv}{kRp.taggedText}(obj) <- value

corpusFreq(obj, ...)

\S4method{corpusFreq}{kRp.taggedText}(obj)

corpusFreq(obj) <- value

\S4method{corpusFreq}{kRp.taggedText}(obj) <- value

corpusCorpFreq(obj, ...)

\S4method{corpusCorpFreq}{kRp.taggedText}(obj)

corpusCorpFreq(obj) <- value

\S4method{corpusCorpFreq}{kRp.taggedText}(obj) <- value

corpusStopwords(obj, ...)

\S4method{corpusStopwords}{kRp.taggedText}(obj)

corpusStopwords(obj) <- value

\S4method{corpusStopwords}{kRp.taggedText}(obj) <- value

\S4method{[}{kRp.taggedText,ANY,ANY,ANY}(x, i, j)

\S4method{[}{kRp.taggedText,ANY,ANY,ANY}(x, i, j) <- value

\S4method{[[}{kRp.taggedText}(x, i)

\S4method{[[}{kRp.taggedText}(x, i) <- value

\S4method{describe}{kRp.taggedText}(obj)

\S4method{describe}{kRp.taggedText}(obj) <- value

\S4method{language}{kRp.taggedText}(obj)

\S4method{language}{kRp.taggedText}(obj) <- value

diffText(obj, value)

\S4method{diffText}{kRp.taggedText}(obj)

diffText(obj) <- value

\S4method{diffText}{kRp.taggedText}(obj) <- value

originalText(obj, value)

\S4method{originalText}{kRp.taggedText}(obj)

is.taggedText(obj)

fixObject(obj, doc_id = NA)

\S4method{fixObject}{kRp.taggedText}(obj, doc_id = NA)

tif_as_tokens_df(tokens)

\S4method{tif_as_tokens_df}{kRp.taggedText}(tokens)
}
\arguments{
\item{obj}{An arbitrary \code{R} object.}

\item{add.desc}{Logical,
      determines whether the \code{desc} column should be re-written with descriptions
for all POS tags.}

\item{doc_id}{Logical (except for \code{fixObject}),
      if \code{TRUE} the \code{doc_id} column will be a factor with the respective value
of the \code{desc} slot, i.\,e.,
      the document ID will be preserved in the data.frame. If used with \code{fixObject}, can be a character string
to set the document ID manually (the default \code{NA} will preserve existing values and not overwrite them).}

\item{value}{The new value to replace the current with.}

\item{feature}{Character string naming the feature to look for. The return value is logical if a single feature
name is given. If \code{feature=NULL}, a character vector is returned,
      naming all features found in the object.}

\item{x}{An object of class \code{kRp.taggedText} or \code{kRp.hyphen}.}

\item{i}{Defines the row selector (\code{[}) or the name to match (\code{[[}).}

\item{j}{Defines the column selector.}

\item{tokens}{An object of class \code{\link[koRpus:kRp.tagged-class]{kRp.tagged}}.}
}
\description{
These methods should be used to get or set values of tagged text objects
generated by koRpus functions like \code{treetag()} or \code{tokenize()}.
}
\details{
\itemize{
  \item{\code{taggedText()} }{returns the \code{tokens} slot.}
  \item{\code{describe()} }{returns the \code{desc} slot.}
  \item{\code{language()} }{returns the \code{lang} slot.}
  \item{\code{[}/\code{[[} }{Can be used as a shortcut to index the results of \code{taggedText()}.}
  \item{\code{fixObject} }{returns the same object upgraded to the object structure of this package version (e.g.,
      new columns, changed names, etc.).}
  \item{\code{hasFeature()} }{returns \code{TRUE} or code{FALSE},
      depending on whether the requested feature is present or not.}
  \item{\code{feature()} }{returns the list entry of the \code{feat_list} slot for the requested feature.}
  \item{\code{corpusReadability()} }{returns the list of \code{kRp.readability} objects.}
  \item{\code{corpusHyphen()} }{returns the list of \code{kRp.hyphen} objects.}
  \item{\code{corpusLexDiv()} }{returns the list of \code{kRp.TTR} objects.}
  \item{\code{corpusFreq()} }{returns the frequency analysis data from the \code{feat_list} slot.}
  \item{\code{corpusCorpFreq()} }{returns the \code{kRp.corp.freq} object of the \code{feat_list} slot.}
  \item{\code{corpusStopwords()} }{returns the number of stopwords found in each text (if analyzed) from the \code{feat_list} slot.}
  \item{\code{tif_as_tokens_df} }{returns the \code{tokens} slot in a TIF[1] compliant format,
      i.e., \code{doc_id} is not a factor but a character vector.}
  \item{\code{originalText()} }{similar to \code{taggedText()},
      but reverts any transformations back to the original text before returning the \code{tokens} slot.
     Only works if the object has the feature \code{diff}.}
  \item{\code{diffText()} }{returns the \code{diff} slot, if present.}
}
}
\examples{
\dontrun{
taggedText(tagged.txt)
}
}
\references{
[1] Text Interchange Formats (\url{https://github.com/ropensci/tif})
}
