% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.BAWL.R
\name{read.BAWL}
\alias{read.BAWL}
\title{Import BAWL-R data}
\usage{
read.BAWL(csv, fileEncoding = NULL)
}
\arguments{
\item{csv}{A character string, path to the BAWL-R in CSV2 format.}

\item{fileEncoding}{A character string naming the encoding of the file, if necessary.}
}
\value{
An object of class \code{\link[koRpus:kRp.corp.freq-class]{kRp.corp.freq}}.
}
\description{
Read the Berlin Affective Word List -- Reloaded (V\"o, Conrad, Kuchinke, Hartfeld,
      Hofmann & Jacobs, 2009; [1]) into a valid object of class
\code{\link[koRpus:kRp.corp.freq-class]{kRp.corp.freq}}.
}
\details{
To use this function,
      you must first export the BAWL-R list into CSV format: Use comma for decimal values and semicolon as value separator
(often referred to as CSV2). Once you have successfully imported the word list,
      you can use the object to perform frequency analysis.
}
\examples{
\dontrun{
bawl.corp <- read.BAWL(
  file.path("~","mydata","valence","BAWL-R.csv")
)

# you can now use query() now to create subsets of the word list,
# e.g., only nound with 5 letters and an valence rating of >= 1
bawl.stimulus <- query(bawl.corp,
  query=list(
    list(wclass="noun"),
    list(lttr=5),
    list("EMO_MEAN"=1, rel="ge")
  )
)
}
}
\references{
V\"o, M. L.-H., Conrad, M., Kuchinke, L., Hartfeld, K., Hofmann, M.F. & Jacobs,
      A.M. (2009).
  The Berlin Affective Word List Reloaded (BAWL-R). \emph{Behavior Research Methods},
      41(2), 534--538.
  doi: 10.3758/BRM.41.2.534

[1] \url{https://www.ewi-psy.fu-berlin.de/psychologie/arbeitsbereiche/allgpsy/Download/BAWL/index.html}
}
\seealso{
\code{\link[koRpus:kRp.corp.freq-class]{kRp.corp.freq}},
  \code{\link[koRpus:query]{query}}
}
\keyword{corpora}
