% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions_readability.R
\name{SMOG}
\alias{SMOG}
\title{Readability: Simple Measure of Gobbledygook (SMOG)}
\usage{
SMOG(txt.file, hyphen = NULL, parameters = c(syll = 3, sqrt = 1.043, fact =
  30, const = 3.1291, sqrt.const = 0), ...)
}
\arguments{
\item{txt.file}{Either an object of class \code{\link[koRpus:kRp.tagged-class]{kRp.tagged}},
      a character vector which must be be
a valid path to a file containing the text to be analyzed,
      or a list of text features. If the latter, calculation
is done by \code{\link[koRpus:readability.num]{readability.num}}.}

\item{hyphen}{An object of class kRp.hyphen. If \code{NULL},
      the text will be hyphenated automatically.}

\item{parameters}{A numeric vector with named magic numbers,
      defining the relevant parameters for the index.}

\item{...}{Further valid options for the main function,
      see \code{\link[koRpus:readability]{readability}} for details.}
}
\value{
An object of class \code{\link[koRpus:kRp.readability-class]{kRp.readability}}.
}
\description{
This is just a convenient wrapper function for \code{\link[koRpus:readability]{readability}}.
}
\details{
This function calculates the Simple Measure of Gobbledygook (SMOG). In contrast to \code{\link[koRpus:readability]{readability}},
which by default calculates all possible indices,
      this function will only calculate the index value.

By default calculates formula D by McLaughlin (1969).
If \code{parameters} is set to \code{SMOG="C"}, formula C will be calculated.
If \code{parameters} is set to \code{SMOG="simple"}, the simplified formula is used, and
if \code{parameters="de"}, the formula adapted to German texts ("Qu", Bamberger & Vanecek,
      1984, p. 78).
}
\examples{
\dontrun{
SMOG(tagged.text)
}
}
\references{
Bamberger, R. & Vanecek,
      E. (1984). \emph{Lesen--Verstehen--Lernen--Schreiben}. Wien: Jugend und Volk.

   McLaughlin,
      G.H. (1969). SMOG grading -- A new readability formula. \emph{Journal of Reading}, 12(8), 639--646.
}
\keyword{readability}
