% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_method_freq.analysis.R
\docType{methods}
\name{freq.analysis}
\alias{freq.analysis}
\alias{freq.analysis,kRp.taggedText-method}
\alias{freq.analysis,character-method}
\title{Analyze word frequencies}
\usage{
freq.analysis(txt.file, ...)

\S4method{freq.analysis}{kRp.taggedText}(txt.file, corp.freq = NULL,
  desc.stat = TRUE, force.lang = NULL, tagger = "kRp.env",
  corp.rm.class = "nonpunct", corp.rm.tag = c(), tfidf = TRUE, ...)

\S4method{freq.analysis}{character}(txt.file, corp.freq = NULL,
  desc.stat = TRUE, force.lang = NULL, tagger = "kRp.env",
  corp.rm.class = "nonpunct", corp.rm.tag = c(), tfidf = TRUE, ...)
}
\arguments{
\item{txt.file}{Either an object of class \code{\link[koRpus:kRp.tagged-class]{kRp.tagged}},
      \code{\link[koRpus:kRp.txt.freq-class]{kRp.txt.freq}},
\code{\link[koRpus:kRp.analysis-class]{kRp.analysis}} or \code{\link[koRpus:kRp.txt.trans-class]{kRp.txt.trans}},
      or a character vector which must
be a valid path to a file containing the text to be analyzed.}

\item{...}{Additional options to be passed through to the function defined with \code{tagger}.}

\item{corp.freq}{An object of class \code{\link[koRpus:kRp.corp.freq-class]{kRp.corp.freq}}.}

\item{desc.stat}{Logical, whether a descriptive statistical analysis should be performed.}

\item{force.lang}{A character string defining the language to be assumed for the text,
      by force.}

\item{tagger}{A character string defining the tokenizer/tagger command you want to use for basic text analysis. Can be omitted if
\code{txt.file} is already of class \code{kRp.tagged-class}. Defaults to \code{"kRp.env"} to get the settings by
\code{\link[koRpus:get.kRp.env]{get.kRp.env}}. Set to \code{"tokenize"} to use \code{\link[koRpus:tokenize]{tokenize}}.}

\item{corp.rm.class}{A character vector with word classes which should be ignored for frequency analysis. The default value
\code{"nonpunct"} has special meaning and will cause the result of
\code{kRp.POS.tags(lang, c("punct","sentc"), list.classes=TRUE)} to be used.}

\item{corp.rm.tag}{A character vector with POS tags which should be ignored for frequency analysis.}

\item{tfidf}{Logical,
      whether the term frequency--inverse document frequency statistic (tf-idf) should be computed. Requires
\code{corp.freq} to provide appropriate idf values for the types in \code{txt.file}. Missing idf values will result in \code{NA}.}
}
\value{
An object of class \code{\link[koRpus:kRp.txt.freq-class]{kRp.txt.freq}}.
}
\description{
The function \code{freq.analysis} analyzes texts regarding frequencies of tokens,
      word classes etc.
}
\details{
The easiest way to see what kinds of analyses are done is probably to look at the slot description of \code{\link[koRpus:kRp.txt.freq-class]{kRp.txt.freq}}.

By default, if the text has yet to be tagged,
      the language definition is queried by calling \code{get.kRp.env(lang=TRUE)} internally.
Or, if \code{txt.file} has already been tagged,
      by default the language definition of that tagged object is read
and used. Set \code{force.lang=get.kRp.env(lang=TRUE)} or to any other valid value,
      if you want to forcibly overwrite this
default behaviour,
      and only then. See \code{\link[koRpus:kRp.POS.tags]{kRp.POS.tags}} for all supported languages.
}
\note{
Prior to \code{koRpus} 0.04-29, this function was named \code{kRp.freq.analysis()}.
   For backwards compatibility there is a wrapper function, but it should be considered
   deprecated.
}
\examples{
\dontrun{
freq.analysis("~/some/text.txt", corp.freq=my.LCC.data)
}
}
\seealso{
\code{\link[koRpus:get.kRp.env]{get.kRp.env}},
      \code{\link[koRpus:kRp.tagged-class]{kRp.tagged}},
   \code{\link[koRpus:kRp.corp.freq-class]{kRp.corp.freq}}
}
\keyword{misc}
