% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.kRp.env.R
\name{get.kRp.env}
\alias{get.kRp.env}
\title{Get koRpus session settings}
\usage{
get.kRp.env(..., errorIfUnset = TRUE)
}
\arguments{
\item{...}{Named parameters to get from the koRpus environment. Valid arguments are:
\describe{
  \item{TT.cmd}{ Logical, whether the set tagger command should be returned.}
  \item{lang}{ Logical, whether the set language should be returned.}
  \item{TT.options}{ Logical,
      whether the set TT.options for \code{treetag} should be returned.}
  \item{hyph.cache.file}{ Logical,
      whether the set hyphenation cache file for \code{hyphen} should be returned.}
  \item{add.desc}{ Logical,
      whether tag descriptions should be added directly to tagged text objects.}
}}

\item{errorIfUnset}{Logical, if \code{TRUE} and the desired property is not set at all,
      the function will fail with an error message.}
}
\value{
A character string or list, possibly including:
 \item{TT.cmd}{Path information for the TreeTagger command}
 \item{lang}{The specified language}
 \item{TT.options}{A list with options for \code{treetag}}
 \item{hyph.cache.file}{The specified hyphenation cache file for \code{hyphen}}
}
\description{
The function \code{get.kRp.env} returns information on your session environment regarding the koRpus package,
      e.g.
where your local TreeTagger installation resides, if it was set before using
\code{\link[koRpus:set.kRp.env]{set.kRp.env}}.
}
\details{
For the most part,
      \code{get.kRp.env} is a convenient wrapper for \code{\link[base:getOption]{getOption}}.
}
\examples{
\dontrun{
set.kRp.env(TT.cmd="~/bin/treetagger/cmd/tree-tagger-german", lang="de")
get.kRp.env(TT.cmd=TRUE)
}
}
\seealso{
\code{\link[koRpus:set.kRp.env]{set.kRp.env}}
}
\keyword{misc}
