% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_method_filterByClass.R,
%   R/02_method_text.transform.R
\name{koRpus-deprecated}
\alias{koRpus-deprecated}
\alias{kRp.filter.wclass}
\alias{kRp.text.transform}
\title{Deprecated functions}
\usage{
kRp.filter.wclass(...)

kRp.text.transform(...)
}
\arguments{
\item{...}{Parameters to be passed to the replacement of the function}

\item{...}{Parameters to be passed to the replacement of the function}
}
\description{
These functions will be removed soon and should no longer ne used.
}
