\docType{package}
\name{koRpus-package}
\alias{koRpus}
\alias{koRpus-package}
\title{The koRpus Package}
\description{
  An R Package for Text Analysis.
}
\details{
  \tabular{ll}{ Package: \tab koRpus\cr Type: \tab
  Package\cr Version: \tab 0.04-40\cr Date: \tab
  2013-04-07\cr Depends: \tab R (>= 2.10.0),methods\cr
  Enhances: \tab rkward\cr Encoding: \tab UTF-8\cr License:
  \tab GPL (>= 3)\cr LazyLoad: \tab yes\cr URL: \tab
  http://reaktanz.de/?c=hacking&s=koRpus\cr }

  A set of tools to analyze texts. Includes, amongst
  others, functions for automatic language detection,
  hyphenation, several indices of lexical diversity (e.g.,
  type token ratio, HD-D/vocd-D, MTLD) and readability
  (e.g., Flesch, SMOG, LIX, Dale-Chall). Basic import
  functions for language corpora are also provided, to
  enable frequency analyses (supports Celex and Leipzig
  Corpora Collection file formats).

  Note: For full functionality a local installation of
  TreeTagger is recommended. Be encouraged to send feedback
  to the author(s)!
}
\author{
  Meik Michalke \email{meik.michalke@hhu.de}, with
  contributions from Earl Brown \email{eabrown@csumb.edu},
  Alberto Mirisola, Alexandre Brulet, Laura Hauser
}
\keyword{package}

