\docType{package}
\name{koRpus-package}
\alias{koRpus-package}
\title{The koRpus Package}
\description{
An R Package for Text Analysis.
}
\details{
\tabular{ll}{ Package: \tab koRpus\cr Type: \tab Package\cr
Version: \tab 0.05-5\cr Date: \tab 2014-03-19\cr Depends:
\tab R (>= 2.10.0),methods\cr Enhances: \tab rkward\cr
Encoding: \tab UTF-8\cr License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr URL: \tab
http://reaktanz.de/?c=hacking&s=koRpus\cr }

A set of tools to analyze texts. Includes, amongst others,
functions for automatic language detection, hyphenation,
several indices of lexical diversity (e.g., type token
ratio, HD-D/vocd-D, MTLD) and readability (e.g., Flesch,
SMOG, LIX, Dale-Chall). Basic import functions for language
corpora are also provided, to enable frequency analyses
(supports Celex and Leipzig Corpora Collection file
formats) and measures like tf-idf.

Note: For full functionality a local installation of
TreeTagger is recommended. koRpus also includes a plugin
for the R GUI and IDE RKWard, providing dialogs for its
basic features. To use them, install RKWard from
http://rkward.sf.net (plugins are detected automatically).
Due to some restrictions on CRAN, the full package sources
are only available from the project homepage.
}
\author{
m.eik michalke, with contributions from Earl Brown, Alberto
Mirisola, Alexandre Brulet, Laura Hauser
}
\keyword{package}

