% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_class_04_kRp.lang.R
\docType{class}
\name{kRp.lang,-class}
\alias{kRp.lang,-class}
\alias{kRp_lang}
\alias{kRp.lang-class}
\title{S4 Class kRp.lang}
\description{
This class is used for objects that are returned by \code{\link[koRpus:guess.lang]{guess.lang}}.
}
\section{Slots}{

\describe{
\item{\code{lang}}{A character string,
      naming the language (by its ISO 639-3 identifier) that was estimated for the analized text in this object.}

\item{\code{lang.name}}{A character string, full name of the estimated language.}

\item{\code{txt}}{A character string containing the analized part of the text.}

\item{\code{txt.full}}{A character string containing the full text.}

\item{\code{udhr}}{A data.frame with full analysis results for each language tried.}
}}

\section{Contructor function}{

Should you need to manually generate objects of this class (which should rarely be the case),
      the contructor function 
\code{kRp_lang(...)} can be used instead of
\code{new("kRp.lang", ...)}.
}

\keyword{classes}
