% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_method_kRp.taggedText.R
\docType{methods}
\name{taggedText}
\alias{taggedText}
\alias{taggedText,kRp.taggedText-method}
\alias{taggedText,-methods}
\alias{taggedText<-}
\alias{taggedText<-,kRp.taggedText-method}
\alias{taggedText<-,-methods}
\alias{[,kRp.taggedText-method}
\alias{[,-methods}
\alias{[,kRp.taggedText,ANY,ANY-method}
\alias{[<-,kRp.taggedText-method}
\alias{[<-,-methods}
\alias{[<-,kRp.taggedText,ANY,ANY,ANY-method}
\alias{[[,kRp.taggedText-method}
\alias{[[,-methods}
\alias{[[,kRp.taggedText,ANY-method}
\alias{[[<-,kRp.taggedText-method}
\alias{[[<-,-methods}
\alias{[[<-,kRp.taggedText,ANY,ANY-method}
\alias{describe}
\alias{describe,kRp.taggedText-method}
\alias{describe,-methods}
\alias{describe<-}
\alias{describe<-,kRp.taggedText-method}
\alias{describe<-,-methods}
\alias{language}
\alias{language,kRp.taggedText-method}
\alias{language,-methods}
\alias{language<-}
\alias{language<-,kRp.taggedText-method}
\alias{language<-,-methods}
\alias{is.taggedText}
\title{Getter/setter methods for koRpus objects}
\usage{
taggedText(obj)

\S4method{taggedText}{kRp.taggedText}(obj)

taggedText(obj) <- value

\S4method{taggedText}{kRp.taggedText}(obj) <- value

\S4method{[}{kRp.taggedText}(x, i, j)

\S4method{[}{kRp.taggedText}(x, i, j) <- value

\S4method{[[}{kRp.taggedText}(x, i)

\S4method{[[}{kRp.taggedText}(x, i) <- value

describe(obj)

\S4method{describe}{kRp.taggedText}(obj)

describe(obj) <- value

\S4method{describe}{kRp.taggedText}(obj) <- value

language(obj)

\S4method{language}{kRp.taggedText}(obj)

language(obj) <- value

\S4method{language}{kRp.taggedText}(obj) <- value

is.taggedText(obj)
}
\arguments{
\item{obj}{An arbitrary \code{R} object.}

\item{value}{The new value to replace the current with.}

\item{x}{An object of class \code{kRp.taggedText} or \code{kRp.hyphen}.}

\item{i}{Defines the row selector (\code{[}) or the name to match (\code{[[}).}

\item{j}{Defines the column selector.}
}
\description{
These methods should be used to get or set values of tagged text objects
generated by koRpus functions like \code{treetag()} or \code{tokenize()}.
}
\details{
\itemize{
  \item{\code{taggedText()} }{returns the \code{TT.res} slot.}
  \item{\code{describe()} }{returns the \code{desc} slot.}
  \item{\code{language()} }{returns the \code{lang} slot.}
  \item{\code{[}/\code{[[} }{Can be used as a shortcut to index the results of \code{taggedText()}.}
}
}
\examples{
\dontrun{
taggedText(tagged.txt)
}
}
