% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_class_03_kRp.txt.freq.R
\docType{class}
\name{kRp.txt.freq,-class}
\alias{kRp.txt.freq,-class}
\alias{kRp.txt.freq-class}
\title{S4 Class kRp.txt.freq}
\description{
This class is used for objects that are returned by \code{\link[koRpus:freq.analysis]{freq.analysis}}.
}
\section{Slots}{

\describe{
\item{\code{lang}}{A character string,
      naming the language that is assumed for the analized text in this object.}

\item{\code{TT.res}}{A data.frame with a version of the fully tagged text (like \code{TT.res} in class \code{koRpus.tagged},
      plus frequency data).}

\item{\code{desc}}{A list with detailed descriptive statistics on the analyzed text.}

\item{\code{freq.analysis}}{A list with information on the word frequencies of the analyzed text.}
}}

\keyword{classes}
