% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_method_pasteText.R
\docType{methods}
\name{pasteText}
\alias{pasteText}
\alias{pasteText,kRp.taggedText-method}
\title{Paste koRpus objects}
\usage{
pasteText(txt, replace = c(hon.kRp = "", hoff.kRp = "\\n\\n", p.kRp =
  "\\n\\n"), ...)

\S4method{pasteText}{kRp.taggedText}(txt, replace = c(hon.kRp = "",
  hoff.kRp = "\\n\\n", p.kRp = "\\n\\n"))
}
\arguments{
\item{txt}{An object of class \code{\link[koRpus:kRp.txt.trans-class]{kRp.txt.trans}},
      \code{\link[koRpus:kRp.tagged-class]{kRp.tagged}},
\code{\link[koRpus:kRp.txt.freq-class]{kRp.txt.freq}} or \code{\link[koRpus:kRp.analysis-class]{kRp.analysis}}.}

\item{replace}{A named character vector to define replacements for \code{koRpus}' internal headline and paragraph tags.}

\item{...}{Additional options, currently unused.}
}
\value{
An atomic character vector.
}
\description{
Paste the text in koRpus objects.
}
\details{
This function takes objects of either class \code{kRp.tagged},
      \code{kRp.txt.freq} or \code{kRp.analysis} and pastes only
the actual text as is.
}
\examples{
\dontrun{
  tagged.text.obj <- freq.analysis("/some/text.txt", corp.freq=my.LCC.data)
  pasteText(tagged.text.obj)
}
}
\keyword{misc}
